:recycle: Refactor #1: 클린 아키텍쳐 구조 적용
