+++ app/src/main/java/com/umc/edison/MainActivity.kt
@@ -1,4 +1,4 @@
-package com.umc.edison.presentation
+package com.umc.edison
 
 import android.os.Bundle
 import androidx.activity.ComponentActivity
@@ -11,7 +11,7 @@ import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.tooling.preview.Preview
-import com.umc.edison.presentation.theme.EdisonTheme
+import com.umc.edison.ui.theme.EdisonTheme
 
 class MainActivity : ComponentActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
