+++ app/src/main/java/com/umc/edison/ui/theme/Color.kt
@@ -1,4 +1,4 @@
-package com.umc.edison.presentation.theme
+package com.umc.edison.ui.theme
 
 import androidx.compose.ui.graphics.Color
