+++ app/src/main/java/com/umc/edison/ui/theme/Type.kt
@@ -1,4 +1,4 @@
-package com.umc.edison.presentation.theme
+package com.umc.edison.ui.theme
 
 import androidx.compose.material3.Typography
 import androidx.compose.ui.text.TextStyle
